Move CmdBase class to common dir
