+++ dvc/command/checkout.py
@@ -1,6 +1,6 @@
 import os
 
-from dvc.command.base import CmdBase, DvcLock
+from dvc.command.common.base import CmdBase, DvcLock
 from dvc.logger import Logger
 from dvc.system import System
