+++ dvc/command/init.py
@@ -1,7 +1,7 @@
 import os
 from pathlib import Path
 
-from dvc.command.base import CmdBase
+from dvc.command.common.base import CmdBase
 from dvc.logger import Logger
 from dvc.config import Config
 from dvc.exceptions import DvcException
