+++ dvc/command/lock.py
@@ -1,4 +1,4 @@
-from dvc.command.base import CmdBase, DvcLock
+from dvc.command.common.base import CmdBase, DvcLock
 from dvc.logger import Logger
 from dvc.state_file import StateFile
