+++ dvc/command/merge.py
@@ -1,12 +1,12 @@
 import os
-import json
 
-from dvc.command.base import CmdBase, DvcLock
+from dvc.command.common.base import CmdBase, DvcLock
 from dvc.config import ConfigI
 from dvc.logger import Logger
 from dvc.state_file import StateFile
 from dvc.path.data_item import DataItem
 from dvc.system import System
+
 
 class CmdMerge(CmdBase):
     def __init__(self, settings):
