+++ dvc/command/show_pipeline.py
@@ -2,7 +2,6 @@ import os
 import networkx as nx
 
 from dvc.command.traverse import Traverse
-from dvc.command.init import CmdInit
 from dvc.logger import Logger
 from dvc.state_file import StateFile
