+++ dvc/command/traverse.py
@@ -1,6 +1,6 @@
 import os
 
-from dvc.command.base import CmdBase, DvcLock
+from dvc.command.common.base import CmdBase, DvcLock
 from dvc.exceptions import DvcException
 from dvc.logger import Logger
 from dvc.path.data_item import DataItemError
